core/pyproject.toml
[project]
name = "ai-avatar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "miumo6329", email = "miumo6329@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.86.0",
    "faster-whisper>=1.2.1",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    "websockets>=16.0",
]

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"
